refactor: Add reverse_vector .cpp/.h files
reverse_vector.h: vector and function declarations
reverse_vector.cpp: vector and function definitions
