+++ pt-1-the-basics/chp-8-technicalities-functions-etc/exercises/reverse_vector/headers/reverse_vector.h
@@ -1,0 +1,6 @@
+#include <vector>
+
+extern std::vector<int> data;
+std::vector<int> reverse_by_value(const std::vector<int> v);
+std::vector<int> reverse_by_reference(std::vector<int>& v);
+void print(const std::vector<int> v);
